I never posted this part, but this is the code to 10 print
This is the code to making a similar 10print picture. I might upload the picture, later but this is code to make a 10 print like that read me in javascript, using p5. The difference is that in the code, instead of it being multicolour, one dash is blue and the other type is red. all you need to do is add randoms. If you plan on making this in the future, I suggest using processing in p5 mode. This is what I use and I find this the easiest, however I'm using it on a MacBook and it may be different for windows users.